Fix issue where user key saved to key of same value
The user key was being saved to a key that was the text of the user key
instead of adzUserKey. This caused the user key to essentially be lost
every time, a new key being generated, and then saved to a completely
different key in the keychain. Everything pulled out of adzUserKey
would always be nil.